Sync new events that occur when "suspended" (#23)
* Add initial attempt to sync events received while suspended

Refs #22

It's a bit kludgey and could use some testing, but is fairly
thorough

* Begin testing the resume-sync logic

Refs #22

We also could/should use this setup to test all the many other
things ConnectionManager does

* Add more tests

Closes #22

There's still plenty more testing we should do, but this covers
the sync-while-suspended thing fine for now. We'll want to do
a live test and make sure the timestamp we used doesn't cause
dups if we received a message while suspended.