Disable application menu on windows
Refs #43

We don't seem to need it for ctrl-c/v to work like on macOS, and
it looks quite awful right now.

I think we'll leave the frame for the time being, however; while
it won't look as nice as on macOS, implementing the close/minimize
buttons is more work than it's worth.