Update javascript dependencies; add package-lock.json
Refs #49

We haven't updated electron-packager yet because there appear to
be some breaking changes we'll have to look into. Non-critical.

We still get a warning on npm install about using a deprecated
tough-cookie package, but since it seems to only be required by
hangupsjs, and that doesn't *actually* use it (sigh) I don't think
it's a problem for now.