Revert updates to cljs and figwheel
They don't like our use of :npm-deps. I think there's a lot of
refactoring we could do of our config here, but I don't have the
energy to do that right now. Ideally we wouldn't need all the bogus
"expect" scripts we have for the repl, and could instead load the
right configs and do the right things from the user ns....